Move MeasurementResult into _typing to fix circular dependency in #1449
