+++ mitiq/tests/test_measurement_result.py
@@ -17,7 +17,7 @@
 import pytest
 
 import numpy as np
-from mitiq.rem.measurement_result import MeasurementResult
+from mitiq._typing import MeasurementResult
 
 
 @pytest.mark.parametrize("asarray", (True, False))
